+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/extensions/jetpack-contact-form/constants.ts
@@ -1,4 +1,4 @@
-export const JETPACK_FORM_AI_COMPOSITION_EXTENSION = 'ai-assistant-form-support';
+export const JETPACK_FORM_AI_COMPOSITION_EXTENSION = 'ai-assistant-form-support' as const;
 
 export const isJetpackFromBlockAiCompositionAvailable =
 	window?.Jetpack_Editor_Initial_State.available_blocks?.[ JETPACK_FORM_AI_COMPOSITION_EXTENSION ]
