+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/hooks/use-ai-feature/index.ts
@@ -3,20 +3,11 @@
  */
 import apiFetch from '@wordpress/api-fetch';
 import { useEffect, useState } from 'react';
-
+/**
+ * Types & constants
+ */
 export type UpgradeTypeProp = 'vip' | 'default';
-
-export type SiteAIAssistantFeatureEndpointResponseProps = {
-	'has-feature': boolean;
-	'is-over-limit': boolean;
-	'requests-count': number;
-	'requests-limit': number;
-	'site-require-upgrade': boolean;
-	'error-message': string;
-	'error-code': string;
-	'is-playground-visible': boolean;
-	'upgrade-type': UpgradeTypeProp;
-};
+import type { SiteAIAssistantFeatureEndpointResponseProps } from '../../../../types';
 
 type AIFeatureProps = {
 	hasFeature: boolean;
